Created a new environments folder for storing environment definition files
Updated the SHTOOLS cmake file to be able to find a conda version

Restructured code to allow it to be compiled more robustly without having to call the build_dependencies.sh script directly when building in a conda environment.

Moved the SHTOOLS build script to the FindSHTOOLS.cmake file

Restructured the cli script so that only the main one is needed, and it will fall back to `swiftest` if `swiftest_caf` does not exist.